cleanup

+++ Collections.subproj/MPWUShortArray.h
@@ -46,7 +46,7 @@ THE POSSIBILITY OF SUCH DAMAGE.
 -(unsigned short*)ushorts;
 -(unsigned)ushortAtIndex:(unsigned)index;
 -(unsigned)lastUshort;
--(unsigned)count;
+-(NSUInteger)count;
 -(void)appendUshorts:(const unsigned short*)newData count:(unsigned)newCount;
 -(void)pushUshort:(unsigned short)newShort;
 -(void)popUshort;
